+++ lib/src/app/controller/theme.dart
@@ -25,7 +25,7 @@ import 'package:flutter/material.dart' show Colors, MaterialColor, ThemeData;
 
 import 'package:weathercast/src/app/model.dart' show WeatherCondition;
 
-import 'package:mvc_application/controller.dart' show AppController;
+import 'package:weathercast/src/app/controller.dart' show AppController;
 
 class ThemeCon extends AppController {
   static ThemeCon _this;