+++ lib/src/app/weather_locations/mvc.dart
@@ -1,8 +1,6 @@
 ///
 /// Copyright (C) 2019 Andrious Solutions
 ///
-/// Original Contributor Felix Angelov of Skokie, Illinois
-///
 /// This program is free software; you can redistribute it and/or
 /// modify it under the terms of the GNU General Public License
 /// as published by the Free Software Foundation; either version 3
@@ -19,25 +17,13 @@
 /// See the License for the specific language governing permissions and
 /// limitations under the License.
 ///
-///          Created  13 Feb 2019
+///          Created  21 Feb 2019
 ///
 ///
 
-import 'package:http/http.dart' show Client;
+export 'package:weathercast/src/app/weather_locations/model/weather_locations.dart';
 
-import 'package:weathercast/src/app/model.dart' show Weather, WeatherApiClient;
+export 'package:weathercast/src/app/weather_locations/view/weather_locations.dart';
 
+export 'package:weathercast/src/app/weather_locations/controller/weather_locations.dart';
 
-class WeatherRepository {
-
-  final WeatherApiClient weatherApiClient = WeatherApiClient(
-    httpClient: Client(),
-  );
-
-  WeatherRepository();
-
-  Future<Weather> getWeather(String city) async {
-    final int locationId = await weatherApiClient.getLocationId(city);
-    return weatherApiClient.fetchWeather(locationId);
-  }
-}