+++ CHANGELOG.md
@@ -7,3 +7,11 @@
 ##                       ThemeCon extends AppController {
 ##                       settings_drawer.dart
 ##                       runApp(MVC(WeatherApp()));
+## [0.3.0] - 2019-02-24: 
+##                       class WeatherApp extends AppController {
+##                       class LocationCon extends ControllerMVC {
+##                       
+##                       
+##                       
+##                       
+##